cloud_functions/hubspot_sync/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.13.0
hubspot-api-client==7.0.0

